+++ src/lib/setupContextUtils.js
@@ -233,6 +233,28 @@ function buildPluginApi(tailwindConfig, context, { variantList, variantMap, offs
           .push([{ sort: offset, layer: 'base' }, rule])
       }
     },
+    /**
+     * @param {string} group
+     * @param {Record<string, string | string[]>} declarations
+     */
+    addDefaults(group, declarations) {
+      const groups = {
+        [`@defaults ${group}`]: declarations,
+      }
+
+      for (let [identifier, rule] of withIdentifiers(groups)) {
+        let prefixedIdentifier = prefixIdentifier(identifier, {})
+        let offset = offsets.base++
+
+        if (!context.candidateRuleMap.has(prefixedIdentifier)) {
+          context.candidateRuleMap.set(prefixedIdentifier, [])
+        }
+
+        context.candidateRuleMap
+          .get(prefixedIdentifier)
+          .push([{ sort: offset, layer: 'defaults' }, rule])
+      }
+    },
     addComponents(components, options) {
       let defaultOptions = {
         respectPrefix: true,
@@ -528,6 +550,7 @@ function registerPlugins(plugins, context) {
   let variantList = []
   let variantMap = new Map()
   let offsets = {
+    defaults: 0n,
     base: 0n,
     components: 0n,
     utilities: 0n,
@@ -555,6 +578,7 @@ function registerPlugins(plugins, context) {
 
   let highestOffset = ((args) => args.reduce((m, e) => (e > m ? e : m)))([
     offsets.base,
+    offsets.defaults,
     offsets.components,
     offsets.utilities,
     offsets.user,
@@ -566,13 +590,14 @@ function registerPlugins(plugins, context) {
   context.arbitraryPropertiesSort = ((1n << reservedBits) << 0n) - 1n
 
   context.layerOrder = {
-    base: (1n << reservedBits) << 0n,
-    components: (1n << reservedBits) << 1n,
-    utilities: (1n << reservedBits) << 2n,
-    user: (1n << reservedBits) << 3n,
+    defaults: (1n << reservedBits) << 0n,
+    base: (1n << reservedBits) << 1n,
+    components: (1n << reservedBits) << 2n,
+    utilities: (1n << reservedBits) << 3n,
+    user: (1n << reservedBits) << 4n,
   }
 
-  reservedBits += 4n
+  reservedBits += 5n
 
   let offset = 0
   context.variantOrder = new Map(